requirements.txt
absl-py==0.12.0
altair==4.1.0
ann-visualizer==2.5
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
blis==0.7.4
branca==0.4.2
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.4
catboost==0.26
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chromedriver-binary==74.0.3729.6.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.3
ExifRead==2.3.2
Flask==2.0.1
flatbuffers==1.12
folium==0.12.1
future==0.18.2
gast==0.4.0
gensim==4.0.1
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
importlib-resources==5.2.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-visualizer==2.4
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.1
mistune==0.8.4
murmurhash==1.0.5
mysql-connector-python==8.0.23
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
preshed==3.0.5
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.4
pydeck==0.6.2
pydot==1.4.2
pydotplus==2.0.2
pyfiglet==0.8.post1
pyglet==1.5.16
Pygments==2.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
ratelim==0.1.6
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smart-open==3.0.0
smmap==4.0.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.4.17
squarify==0.4.3
srsly==2.4.1
streamlit==0.82.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-metadata==1.1.0
termcolor==1.1.0
terminado==0.9.5
testpath==0.5.0
textblob==0.15.3
thinc==8.0.3
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
tweety==0.1.6
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
vaderSentiment==3.3.2
validators==0.18.2
vega-datasets==0.9.0
wasabi==0.8.2
watchdog==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
xmlx==2.0.0
zipp==3.4.1
